Refactor GQS average calculation by replacing inline symbol interpretation with a dedicated Interpret method

+++ day-18/src/Symbol.cs
@@ -1,6 +1,6 @@
 namespace GlacialQuantifierSystem;
 
-public record Symbol(int Value)
+public record Symbol(double Value)
 {
     public static Symbol Parse(char input)
         => new(
